Massive compile and running overhaul

+++ database/migrations/2014_10_12_200000_create_problems_table.php
@@ -16,8 +16,8 @@ class CreateProblemsTable extends Migration
             $table->increments('id');
             $table->string('name');
             $table->text('statement');
-            $table->unsignedSmallInteger('time_limit')->default(0);
-            $table->unsignedSmallInteger('memory_limit')->default(0);
+            $table->unsignedInteger('time_limit')->default(0);
+            $table->unsignedInteger('memory_limit')->default(0);
             $table->integer('author_id')->unsigned();
 
             $table->foreign('author_id')->references('id')->on('users');
